Chore: info-item 파일 추가, InfoItem 이동
